backend/modules/token_monitor/requirements.txt
# Token Monitor Module Requirements
# 새 프로젝트에서 이 모듈을 사용할 때 필요한 패키지들

# 필수 패키지
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0

# 선택적 패키지 (웹훅 알림용)
# slack-sdk>=3.0.0  # Slack 알림 사용 시

# 선택적 패키지 (데이터 압축용)
# gzip  # Python 내장 모듈

# 선택적 패키지 (웹 프레임워크)
# fastapi>=0.100.0  # FastAPI 사용 시
# uvicorn>=0.20.0   # FastAPI 서버 실행 시
